requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
audio-recorder-streamlit==0.0.9
AudioRecorder==0.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
-e git+https://github.com/Oyeon/ChatGPT-application-python.git@1f17ba76543bc984d9074f9907b32480c3eacf6e#egg=chatgpt_v01
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
entrypoints==0.4
faiss-cpu==1.8.0.post1
fastapi==0.111.1
fastapi-cli==0.0.4
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
gTTS==2.5.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib-metadata==6.11.0
instagrapi==2.1.2
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langsmith==0.1.93
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==2.0.0
openai==1.35.13
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.27.2
pyarrow==16.1.0
PyAudio==0.2.14
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
streamlit==1.36.0
streamlit-audiorecorder==0.0.5
tenacity==8.5.0
tiktoken==0.7.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
validators==0.33.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
youtube-transcript-api==0.6.2
zipp==3.19.2
